adding the folder structure
