+++ console/.env
@@ -1,0 +1,1 @@
+BASE_URL = 'http://127.0.0.1:5000'
