+++ console/cart.py
@@ -1,0 +1,14 @@
+import requests
+import os
+from dotenv import load_dotenv
+load_dotenv()
+
+BASE_URL = os.getenv('BASE_URL')
+
+
+def view_Cart():
+    return
+
+
+def add_To_Cart():
+    return
