+++ console/product_review.py
@@ -1,0 +1,9 @@
+import requests
+import os
+from dotenv import load_dotenv
+load_dotenv()
+
+BASE_URL = os.getenv('BASE_URL')
+
+def add_Review():
+    return
